Reset properties hook
